게시글 차단, 차단해제 기능 추가

+++ popeye-backend/src/main/java/popeye/popeyebackend/content/repository/ContentBanRepository.java
@@ -3,5 +3,5 @@ package popeye.popeyebackend.content.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import popeye.popeyebackend.content.domain.ContentBan;
 
-public interface ContentBanRepository extends JpaRepository<ContentBan, Integer> {
+public interface ContentBanRepository extends JpaRepository<ContentBan, Long> {
 }
